+++ backend/src/data/avg.py
